refactor: move filter generator logic into individual files
- Move most generation logic out of `filter_generator.go`
- Use interfaces for the filter generators to reduce code complexity
- Some file and package renaming

Testing done: All pre-existing unit tests pass without any modifications to inputs/outputs.
