+++ src/go/configgenerator/filtergen/test_lib.go
@@ -1,0 +1,7 @@
+package filtergen
+
+const (
+	testProjectName = "bookstore.endpoints.project123.cloud.goog"
+	testApiName     = "endpoints.examples.bookstore.Bookstore"
+	testConfigID    = "2019-03-02r0"
+)
